+++ test/dp_exchange/coinbase/money_surface_test.exs
@@ -35,7 +35,7 @@ defmodule DpExchange.Coinbase.MoneySurfaceTest do
 
   @credentials %{
     api_key: "organizations/x/apiKeys/y",
-    api_secret: "-----BEGIN EC PRIVATE KEY-----"
+    api_secret: "dGVzdC1zZWNyZXQtdGhpcnR5LXR3by1ieXRlcyEhISE="
   }
 
   defp responding(body) do