+++ test/dp_exchange/coinbase/portfolios_test.exs
@@ -38,7 +38,7 @@ defmodule DpExchange.Coinbase.PortfoliosTest do
 
   @credentials %{
     api_key: "organizations/x/apiKeys/y",
-    api_secret: "-----BEGIN EC PRIVATE KEY-----"
+    api_secret: "dGVzdC1zZWNyZXQtdGhpcnR5LXR3by1ieXRlcyEhISE="
   }
 
   defp responding(body) do